pr feedback. refactor entity query url generation.
